现在可以正确计算心率和呼吸率

+++ HARDWARE/DMA.h
@@ -1,11 +1,11 @@
 #ifndef _DMA_H
 #define _DMA_H
-
+#include "stm32f10x.h"
 // 宏定义
 #define ADC_NUM 1200 // 30秒数据，40Hz采样率
 // 外部变量声明
-extern uint16_t ADC_Data1[2]; // ADC数据缓冲区，改为数组
-extern uint16_t p1[ADC_NUM];  // 温度数据缓冲区
+extern uint32_t ADC_Data1[2]; // ADC数据缓冲区，改为数组
+extern float p1[ADC_NUM];	  // 数据缓冲区
 extern uint16_t index1;		  // 数据索引
 extern uint8_t do_flag;		  // 数据传输完成标志
